requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
rpds-py==0.9.2
ruff==0.0.282
sqlparse==0.4.4
uritemplate==4.1.1
